Add VB6StrPtr and use in mem calls.
Add VB6StrPtr to allow better manipulation of strings and start using it
in LunaLua's mem assignment calls. Not yet used in properties, but will
be later.

Also, stop duplicating so much code across different versions of the mem
call, when you can reference the global one in all of them.